Refactor NewRegistrationProof into a parent NewRegistration component that builds the Proving/Submitting component based on proving mode the user selects.
