+++ client/src/components/RegistrationForm/index.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components/macro'
 
 import { ExistingRegistration } from "./ExistingRegistration";
-import { NewRegistrationProof } from "./NewRegistrationProof";
-import { NewRegistrationSubmit } from "./NewRegistrationSubmit";
+import { NewRegistration } from "./NewRegistration";
 import AccountContext from '../../contexts/Account/AccountContext';
 
  
@@ -18,6 +17,9 @@ export const RegistrationForm: React.FC = () => {
    */
   const [isNewRegistration, setIsNewRegistration] = useState<boolean>(false);
 
+  /*
+   * Handlers
+   */
   const handleUpdateClick = () => {
     setIsNewRegistration(true);
   }
@@ -26,42 +28,22 @@ export const RegistrationForm: React.FC = () => {
     setIsNewRegistration(false);
   }
 
-  // ----- transaction state -----
-  const [submitOrderPublicSignals, setSubmitOrderPublicSignals] = useState<string>('');
-  const [submitOrderProof, setSubmitOrderProof] = useState<string>('');
-
   /*
     Component
   */
   return (
     <Wrapper>
-      <Column>
-        <Content>
-          {!isNewRegistration ? (
-            <ExistingRegistration
-              loggedInWalletAddress={ethereumAddress}
-              handleNewRegistrationClick={handleUpdateClick}
-            />
-          ) : (
-            <NewRegistrationContainer>
-              <Column>
-                <NewRegistrationProof
-                  loggedInWalletAddress={'0x123'}
-                  setSubmitOrderProof={setSubmitOrderProof}
-                  setSubmitOrderPublicSignals={setSubmitOrderPublicSignals}
-                  handleBackClick={handleBackClick}
-                />
-              </Column>
-              <Column>
-                <NewRegistrationSubmit
-                  proof={submitOrderProof}
-                  publicSignals={submitOrderPublicSignals}
-                />
-              </Column>
-            </NewRegistrationContainer>
-          )}
-        </Content>
-      </Column>
+      {!isNewRegistration ? (
+        <ExistingRegistration
+          loggedInWalletAddress={ethereumAddress}
+          handleNewRegistrationClick={handleUpdateClick}
+        />
+      ) : (
+        <NewRegistration
+          loggedInWalletAddress={ethereumAddress}
+          handleBackClick={handleBackClick}
+        />
+      )}
     </Wrapper>
   );
 };
@@ -70,24 +52,4 @@ const Wrapper = styled.div`
   max-width: 660px;
   width: 100%;
   padding-top: 1.5rem;
-`;
-
-const Column = styled.div`
-  gap: 1rem;
-  align-self: flex-start;
-  border-radius: 16px;
-  justify-content: center;
-`;
-
-const Content = styled.div`
-  gap: 1rem;
-  align-self: flex-start;
-`;
-
-const NewRegistrationContainer = styled.div`
-  display: grid;
-  padding: 1.5rem;
-  background-color: #0D111C;
-  border-radius: 16px;
-  border: 1px solid rgba(255, 255, 255, 0.2);
 `;
